Multi-plane clarifications:
- Clearly indicate in fnOfxExtensions.h that
kFnOfxImageEffectPropComponentsPresent property and
kFnOfxImageEffectActionGetClipComponents action should pass PLANES and
not components string as it was originally stated, otherwise this
cannot work with non-paired planes such as the ones defined in
ofxNatron.h

- ofxNatron.h : add a kNatronOfxExtraCreatedPlanes property on the
image effect instance, indicating the user-created planes. This can be
used for instance, in an output choice menu used to allow the user to
select the output plane.

- Remove the kNatronOfxParamPropChoiceHostCanAddOptions property
altogether which does not serve any real purpose